requirements.txt
Flask==1.1.2
python-dotenv==0.13.0
wallstreet==0.3
pymongo==3.10.1
dnspython==1.16.0
pytz==2019.3